handwriting-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
Pillow==11.0.0
